requirements.txt
streamlit
tensorflow
numpy
Pillow
opencv-python-headless # or just opencv-python
pandas
scikit_learn